requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.2
fastapi-users[sqlalchemy]>=12.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cloudinary>=1.36.0
python-multipart>=0.0.6
alembic>=1.12.1
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.1 